requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
pyyaml
qrcode[pil]>=7.4.2
Flask-Limiter
APScheduler